+++ test/remix/client-entry.test.ts
@@ -46,7 +46,8 @@ describe('initializeSentryOnEntryClient', () => {
           })],
 
           replaysSessionSampleRate: 0.1,
-          replaysOnErrorSampleRate: 1
+          replaysOnErrorSampleRate: 1,
+          sendDefaultPii: true
       })"
     `);
   });
@@ -80,7 +81,8 @@ describe('initializeSentryOnEntryClient', () => {
           })],
 
           replaysSessionSampleRate: 0.1,
-          replaysOnErrorSampleRate: 1
+          replaysOnErrorSampleRate: 1,
+          sendDefaultPii: true
       })"
     `);
   });
@@ -120,7 +122,9 @@ describe('initializeSentryOnEntryClient', () => {
             useEffect,
             useLocation,
             useMatches
-          })]
+          })],
+
+          sendDefaultPii: true
       })"
     `);
   });
@@ -188,7 +192,9 @@ describe('initializeSentryOnEntryClient', () => {
             useEffect,
             useLocation,
             useMatches
-          })]
+          })],
+
+          sendDefaultPii: true
       })"
     `);
   });